\name{runShinyMP}
\alias{runShinyMP}
\title{
  Run the Shiny Application that incorporates the PT.41.t15.tmb management proceedure
}
\description{
  Calling \code{runShinyMP} opens a web browser implemented graphical user interface to evaluate the MP.
}
\usage{
  runShinyMP()
}
\arguments{
}
\details{
  Calling \code{runShinyMP} opens a web browser implemented graphical user interface to evaluate the MP.
}
\value{
  This function does not return a value
}
\author{
  \packageAuthor{PT41F.t15.tmb.MP}

  Maintainer: \packageMaintainer{PT41F.t15.tmb.MP}
}

\seealso{
  \code{\link{PT41F.t15.tmb.MP-package}}
}

\note{
  see \code{\link{PT41F.t15.tmb.MP-package}} for examples of use.
}
